+++ docs/changelog.rst
@@ -2,6 +2,7 @@ Changelog
 ---------
 Unreleased
 ^^^^^^^^^^
+- Updated DFXPReader timestamp validation according to TTML time expression specs
 - Updated flashing cues validation for SCCReader to raise a CaptionReadTimingError
 - Fixed SCC translator not recognising special and extended characters
 - Raise CaptionReadTimingError for missing 'start' on SAMIReader