+++ docs/changelog.rst
@@ -4,6 +4,10 @@ Unreleased
 ^^^^^^^^^^
 - Updated Size.from_string() to accept 0 size without measuring unit
 - Replaced ValueError with CaptionReadSyntaxError for invalid sizes passed to Size.from_string()
+- Updated DFXPReader timestamp validation according to TTML time expression specs
+- Updated flashing cues validation for SCCReader to raise a CaptionReadTimingError
+- Fixed SCC translator not recognising special and extended characters
+- Raise CaptionReadTimingError for missing 'start' on SAMIReader
 
 2.0.5
 ^^^^^